feat(studies): complete look-ahead bias study and pytest suite
- Add studies/lookahead_bias/run_lookahead_study.py with:
  - --cutoff-date flag accepting multiple ISO dates
  - Data sliced to ±600/200 days around each cutoff for speed
  - _find_cutoff_idx() to map calendar dates to datearray indices
  - Two-line ORIG/PATCH output, windowed to 3 pre + CUT + 3 post months
  - --months-pre / --months-post CLI args (default 3)

- Update tests/test_lookahead_bias.py:
  - _run_both_pipelines() slices to _PRE_DAYS=600/_POST_DAYS=200
  - Reduces 3-model pytest runtime from ~4.5 min to ~20 s
  - All three models pass: no look-ahead bias detected

- Fix functions/ta/channels.py: add .astype(int) to periods array
  in all 4 channel functions (float slice index TypeError in NumPy ≥1.24)

- Add plans/stock_selection_testing_plan.md with phase status table
- Add session summaries for 2025-02-13 and 2026-02-23

+++ functions/ta/channels.py
@@ -43,7 +43,10 @@ def percentileChannel(x: NDArray[np.floating], minperiod: int, maxperiod: int,
         - For early data points with insufficient history, uses available data
         - Useful for identifying support and resistance levels
     """
-    periods = np.arange(minperiod, maxperiod, incperiod)
+    # Cast to int: lookback periods must be whole trading days; float
+    # increments passed by callers (e.g. (max-min)/7.) would otherwise
+    # produce float slice indices and raise TypeError.
+    periods = np.arange(minperiod, maxperiod, incperiod).astype(int)
     minchannel = np.zeros(len(x), dtype=float)
     maxchannel = np.zeros(len(x), dtype=float)
     for i in range(len(x)):
@@ -95,7 +98,8 @@ def percentileChannel_2D(x: NDArray[np.floating], minperiod: int, maxperiod: int
         - More efficient than calling percentileChannel in a loop
     """
     print(" ... inside percentileChannel_2D ...  x min,mean,max = ", x.min(), x.mean(), x.max())
-    periods = np.arange(minperiod, maxperiod, incperiod)
+    # Cast to int: lookback periods must be whole trading days.
+    periods = np.arange(minperiod, maxperiod, incperiod).astype(int)
     minchannel = np.zeros((x.shape[0], x.shape[1]), dtype=float)
     maxchannel = np.zeros((x.shape[0], x.shape[1]), dtype=float)
     for i in range(x.shape[1]):
@@ -148,7 +152,8 @@ def dpgchannel(x: NDArray[np.floating], minperiod: int, maxperiod: int,
         - For early data points with insufficient history, uses available data
         - Named 'dpg' for Donald P. Galanis (original developer)
     """
-    periods = np.arange(minperiod, maxperiod, incperiod)
+    # Cast to int: lookback periods must be whole trading days.
+    periods = np.arange(minperiod, maxperiod, incperiod).astype(int)
     minchannel = np.zeros(len(x), dtype=float)
     maxchannel = np.zeros(len(x), dtype=float)
     for i in range(len(x)):
@@ -197,7 +202,8 @@ def dpgchannel_2D(x: NDArray[np.floating], minperiod: int, maxperiod: int,
         - For early data points with insufficient history, uses available data
         - Named 'dpg' for Donald P. Galanis (original developer)
     """
-    periods = np.arange(minperiod, maxperiod, incperiod)
+    # Cast to int: lookback periods must be whole trading days.
+    periods = np.arange(minperiod, maxperiod, incperiod).astype(int)
     minchannel = np.zeros((x.shape[0], x.shape[1]), dtype=float)
     maxchannel = np.zeros((x.shape[0], x.shape[1]), dtype=float)
     for i in range(x.shape[1]):
